day_03: static lib

+++ C/Day 03/lib/phoenix/my_strcpy.c
@@ -9,7 +9,7 @@ char *my_strcpy(char *dest, char const *src)
         return NULL;
     }
 
-    for (int i = 0; i < strlen(src); i++)
+    for (size_t i = 0; i < strlen(src); i++)
         dest[i] = src[i];
 
     dest[strlen(src)] = '\0';
