+++ C/Day 03/lib/phoenix/my_strncmp.c
@@ -3,14 +3,14 @@
 #include <ctype.h>
 #include <string.h>
 
-int my_strncmp(char const *s1, char const *s2, int n)
+int my_strncmp(char const *s1, char const *s2, size_t n)
 {
     if (s1 == NULL || s2 == NULL)
     {
         return (s1 == s2) ? 0 : (s1 == NULL) ? -1
                                              : 1;
     }
-    for (int i = 0; i < n; i++)
+    for (size_t i = 0; i < n; i++)
     {
         if (s1[i] == '\0' || s2[i] == '\0')
         {
